stock/dynamic_fetch/requirements.txt
praw
python-dotenv
datasets
requests
beautifulsoup4
psycopg2==2.9.9
psycopg2-binary==2.9.9
# then run python -m spacy download en_core_web_sm
